Add token_auth login required to endpoints

+++ app/blueprints/api/routes.py
@@ -1,5 +1,5 @@
 from . import bp as api
-from .auth import basic_auth
+from .auth import basic_auth, token_auth
 from flask import jsonify, request
 from app.models import Post, User
 
@@ -48,35 +48,44 @@ def get_post(post_id):
 
 
 @api.route('/posts', methods=['POST'])
+@token_auth.login_required
 def create_post():
     if not request.is_json:
         return jsonify({'error': 'Your request content-type must be application/json'}), 400
     # Get the data from request body
     data = request.json
     # Check to make sure all required fields are present
-    for field in ['title', 'body', 'user_id']:
+    for field in ['title', 'body']:
         if field not in data:
             # if not return a 400 response with error
             return jsonify({'error': f'{field} must be in request body'}), 400
     # Get fields from data dict
     title = data.get('title')
     body = data.get('body')
-    user_id = data.get('user_id')
+    user_id = token_auth.current_user().id
     # Add new post to database with request body info
     new_post = Post(title=title, body=body, user_id=user_id)
     return jsonify(new_post.to_dict()), 201
 
 
 @api.route('/posts/<post_id>', methods=['PUT'])
+@token_auth.login_required
 def update_post(post_id):
     post_to_edit = Post.query.get_or_404(post_id)
+    current_user = token_auth.current_user()
+    if current_user != post_to_edit.author:
+        return jsonify({'error': 'You do not have authorization to edit this post'}), 403
     if not request.is_json:
         return jsonify({'error': 'Your request content-type must be application/json'}), 400
     post_to_edit.update(**request.json)
     return jsonify(post_to_edit.to_dict())
 
 @api.route('/posts/<post_id>', methods=['DELETE'])
+@token_auth.login_required
 def delete_post(post_id):
     post_to_delete = Post.query.get_or_404(post_id)
+    current_user = token_auth.current_user()
+    if current_user != post_to_delete.author:
+        return jsonify({'error': 'You do not have authorization to delete this post'}), 403
     post_to_delete.delete()
     return jsonify({'message': 'You have successfully deleted the post'})
